working on documentation with swagger

+++ entrega-final/src/dao/user.dao.js
@@ -1,22 +1,43 @@
 import { userModel } from "../model/user.model.js";
 import { createHash, verifyPassword } from "../utils/hashPassword.js";
+import { CustomError } from "../utils/errors/customError.utils.js";
+import { errors } from "../utils/errors/errors.js";
 
 export class UserDao {
   static async getAll() {
-    return await userModel.find();
+    const users = await userModel.find();
+
+    return users ? users : "There are no users currently";
   }
   static async getById({ userId }) {
-    return await userModel.findById(userId);
+    try {
+      if (userId.length !== 24) {
+        return CustomError.newError(errors.error.wrongUserId);
+      }
+
+      const user = await userModel.findById(userId);
+
+      if (!user) return CustomError.newError(errors.notFound);
+
+      return user;
+    } catch (error) {
+      console.error(error.reason);
+      throw error;
+    }
   }
   static async update({ data }) {
     const { id, body, elementsToChange } = data;
 
     try {
-      const user = await userModel.findById(id);
+      let user = await userModel.findById(id);
 
       if (!user) {
         const result = { message: "user not found" };
         return result;
+      }
+
+      if (elementsToChange.length < 0) {
+        return CustomError.newError(errors.error.empty);
       }
 
       if (elementsToChange.includes("password")) {
@@ -45,43 +66,38 @@ export class UserDao {
         }
       });
 
-      if (elementsToChange.includes("cart")) {
+      if (
+        elementsToChange.includes("cart") ||
+        elementsToChange.includes("id")
+      ) {
         delete updatesToUser.cart;
+        return CustomError.newError(errors.error.changeNotAllowed);
       }
 
       if (Object.keys(updatesToUser).length > 0) {
-        await userModel.findByIdAndUpdate(id, updatesToUser, { new: true });
+        user = await userModel.findByIdAndUpdate(id, updatesToUser, {
+          new: true,
+        });
       }
 
       const result = {
-        message: `${
-          elementsToChange.length > 0
-            ? `Elements changed, if cart was included, it was not modified`
-            : `No elements were sent to modify`
-        }`,
+        message: "Elements changed",
         user,
       };
       return result;
     } catch (error) {
-      if (error.messageFormat === undefined) {
-        return `User not found`;
-      }
-
-      console.error(error);
+      throw error;
     }
   }
   static async delete({ userId }) {
     try {
       const user = await userModel.findById(userId);
 
+      if (!user) return CustomError.newError(errors.notFound);
+
       await user.deleteOne();
       return { message: `user ${user.first_name} deleted` };
     } catch (error) {
-      if (error.messageFormat === undefined) {
-        return `User not found`;
-      }
-      console.log(error.messageFormat);
-
       console.error(error);
     }
   }
